Re-implement FhirPatientService in terms of FhirResourceHandlers
