+++ api/src/main/java/org/openmrs/module/fhir2/api/FhirPatientService.java
@@ -11,14 +11,9 @@ package org.openmrs.module.fhir2.api;
 
 import javax.annotation.Nonnull;
 
-import java.util.Collection;
-import java.util.List;
-
 import ca.uhn.fhir.rest.api.server.IBundleProvider;
 import ca.uhn.fhir.rest.param.TokenParam;
-import org.hl7.fhir.r4.model.Identifier;
 import org.hl7.fhir.r4.model.Patient;
-import org.openmrs.PatientIdentifierType;
 import org.openmrs.module.fhir2.api.search.param.OpenmrsPatientSearchParams;
 import org.openmrs.module.fhir2.api.search.param.PatientSearchParams;
 
@@ -26,12 +21,6 @@ public interface FhirPatientService extends FhirService<Patient> {
 	
 	@Override
 	Patient get(@Nonnull String uuid);
-	
-	List<Patient> getPatientsByIds(@Nonnull Collection<Integer> ids);
-	
-	Patient getById(@Nonnull Integer id);
-	
-	PatientIdentifierType getPatientIdentifierTypeByIdentifier(Identifier identifier);
 	
 	IBundleProvider searchForPatients(PatientSearchParams patientSearchParams);
