+++ api/src/main/java/org/openmrs/module/fhir2/api/dao/FhirPatientDao.java
@@ -15,7 +15,6 @@ import java.util.Collection;
 import java.util.List;
 
 import org.openmrs.Patient;
-import org.openmrs.PatientIdentifierType;
 import org.openmrs.annotation.Authorized;
 import org.openmrs.module.fhir2.api.search.param.SearchParameterMap;
 import org.openmrs.util.PrivilegeConstants;
@@ -35,9 +34,6 @@ public interface FhirPatientDao extends FhirDao<Patient> {
 	
 	@Authorized(PrivilegeConstants.GET_PATIENTS)
 	List<Patient> getPatientsByIds(@Nonnull Collection<Integer> ids);
-	
-	@Authorized(PrivilegeConstants.GET_PATIENT_IDENTIFIERS)
-	PatientIdentifierType getPatientIdentifierTypeByNameOrUuid(String name, String uuid);
 	
 	@Override
 	@Authorized(PrivilegeConstants.GET_PATIENTS)
