+++ api/src/main/java/org/openmrs/module/fhir2/api/dao/FhirPatientIdentifierSystemDao.java
@@ -26,6 +26,18 @@ public interface FhirPatientIdentifierSystemDao extends FhirDaoAop {
 	@Authorized(PrivilegeConstants.GET_IDENTIFIER_TYPES)
 	PatientIdentifierType getPatientIdentifierTypeByUrl(String url);
 	
+	/**
+	 * Looks up a {@link PatientIdentifierType} by name or uuid. A non-retired type matching
+	 * {@code name} or any type matching {@code uuid} qualifies; when {@code uuid} is supplied it wins
+	 * over a name match.
+	 *
+	 * @param name the identifier type name to match, ignoring retired types
+	 * @param uuid the identifier type uuid to match
+	 * @return the matching identifier type, or {@code null} if none matches
+	 */
+	@Authorized(PrivilegeConstants.GET_PATIENT_IDENTIFIERS)
+	PatientIdentifierType getPatientIdentifierTypeByNameOrUuid(String name, String uuid);
+	
 	@Authorized(PrivilegeConstants.GET_IDENTIFIER_TYPES)
 	Optional<FhirPatientIdentifierSystem> getFhirPatientIdentifierSystem(@Nonnull PatientIdentifierType identifierType);
